backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic[email]==2.6.1
httpx==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.10.4
chromadb==0.4.24
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-text-splitters==0.0.1
langchain-chroma==0.0.5
langchain-ollama==0.0.4
langchain-google-genai==1.0.0
neo4j==5.18.0
sqlalchemy==2.0.27
python-docx==1.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
google-generativeai==0.3.2
openpyxl==3.1.2
pandas==2.2.0
tabula-py==2.9.0
camelot-py[cv]==0.11.0
easyocr==1.7.1
asyncpg==0.29.0
redis==5.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
